requirements.txt
fastapi
uvicorn
google-genai
python-multipart
tavily-python
httpx
python-dotenv
soccerdata
xgboost
pandas
numpy
tavily-python
httpx
python-dotenv
scikit-learn
requests
beautifulsoup4
rich
